+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -24,6 +24,12 @@ Notes](../../RELEASENOTES.md).
   re-encoding on every OTLP metric export.
   ([#7307](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7307))
 
+* Fixed the OTLP/HTTP exporter silently dropping data when an export timed out.
+  ([#7455](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7455))
+
+* Fixed the OTLP/gRPC exporter logging incorrectly when an export timed out.
+  ([#7455](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7455))
+
 ## 1.16.0
 
 Released 2026-Jun-10